PLAN DE LYON

+++ Projet Piscine ING2/sommet.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <string>
 
-Sommet::Sommet(int indice, std::string nom, int x, int y) // constructeur
+Sommet::Sommet(int indice, std::string nom, double x, double y) // constructeur
     :m_nom{nom},m_indice{indice},m_x{x},m_y{y}
 {
 
@@ -22,12 +22,12 @@ void Sommet::afficher(Svgfile &svgout, std::vector <int> classement, int nomin)
     svgout.addText(m_x*100-15, m_y*100-10, m_indice, "green");
 }
 
-int Sommet::GetX()
+double Sommet::GetX()
 {
     return m_x;
 }
 
-int Sommet::GetY()
+double Sommet::GetY()
 {
     return m_y;
 }
